+++ src/bootstrap.js
@@ -9,6 +9,7 @@ async function init(){
   await loadData();
   wireEvents();
   syncSettingsUI();
+  if(typeof initOnboarding === 'function') initOnboarding();
   if(typeof initReferenceLibrary === 'function') initReferenceLibrary();
   setLang(state.lang);
   if(typeof initRouter === 'function') initRouter(); else showView('listView');